backend/requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Supabase
supabase>=2.0.0,<2.20.0

# Data validation & config
pydantic>=2.0.0
pydantic-settings>=2.0.0

# JWT verification
PyJWT>=2.0.0
cryptography>=42.0.0

# Async HTTP (push notifications, test client)
httpx>=0.27.0

# Background task scheduler
apscheduler>=3.10.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Linting
ruff>=0.9.0
